requirements.txt
langchain-community>=0.0.28
pypdf>=3.17.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
python-dotenv>=1.0.0
openai>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0